Set the tag of Docker base image: Redis to 6-alpine
Redis 6.0.4 is a little bit dated, there are some security
updates in the futures versions.

Set the image tag to 6-alpine, instead of 6.0.15-alpine or 6.2.5-alpine,
so that we don't need to manually follow the minor and patch versions
manually, will be more convenient.

(Could use with `--pull` parameter when building the image, so that
Docker will always check the latest 6-alpine)